+++ spec/integration/resource/api200/id_pool/update_spec.rb
@@ -1,0 +1,6 @@
+require 'spec_helper'
+
+RSpec.describe OneviewSDK::IDPool, integration: true, type: UPDATE do
+  let(:current_client) { $client }
+  include_examples 'IDPoolUpdateExample', 'integration context'
+end